Security Updated before goto application online

+++ admin/includes/admin_operation/user_ops/delete_user.php
@@ -1,16 +1,23 @@
-<?php 
+<?php
 ///Code...
-    if(isset($_GET['deleteUserId'])){
-        $deleteUserId = $_GET['deleteUserId'];
+if (isset($_GET['deleteUserId'])) {
+    if (isset($_SESSION['usertype'])) {
+        if ($_SESSION['usertype'] == 'admin') {
+            $deleteUserId = $_GET['deleteUserId'];
 
-        //Make routing buffering page when click on delete item..
-        header("Location: users.php");
+            //Make routing buffering page when click on delete item..
+            header("Location: users.php");
 
-        $query = "DELETE FROM `users` WHERE `users_id` = {$deleteUserId}";
-        $delete_query = mysqli_query($connection, $query);
+            $query = "DELETE FROM `users` WHERE `users_id` = {$deleteUserId}";
+            $delete_query = mysqli_query($connection, $query);
 
-        ///Checking query for delete users query...
-        if(!$delete_query){
-            die("Get ERR! when try to make query Delete Users from View All Users ".mysqli_error($connection));
+            ///Checking query for delete users query...
+            if (!$delete_query) {
+                die("Get ERR! when try to make query Delete Users from View All Users " . mysqli_error($connection));
+            }
+
+        }else{
+            echo "You are Not an Admin to delete user!";
         }
     }
+}
